ゲストのFacilityIDを検証するMiddleWareの追加.リポジトリのFactory化.FacilityListのゲスト対応. getUserInfoのゲスト対応

+++ backend/api/constants/role.go
@@ -9,5 +9,5 @@ const (
 )
 
 const (
-	GuestID = -1 // GUESTログインで使用するユーザーID
+	GuestID = int32(-1) // GUESTログインで使用するユーザーID
 )
